added date library option
