+++ modules/swagger-codegen/src/main/java/io/swagger/codegen/languages/JaxRSServerCodegen.java
@@ -9,11 +9,9 @@ import java.io.File;
 import java.util.*;
 
 public class JaxRSServerCodegen extends JavaClientCodegen implements CodegenConfig {
-    protected String dateLibrary = "default";
     protected String title = "Swagger Server";
     protected String implFolder = "src/main/java";
 
-    public static final String DATE_LIBRARY = "dateLibrary";
     public JaxRSServerCodegen() {
         super();
 
@@ -40,14 +38,6 @@ public class JaxRSServerCodegen extends JavaClientCodegen implements CodegenConf
                 break;
             }
         }
-
-        CliOption dateLibrary = new CliOption(DATE_LIBRARY, "Option. Date library to use");
-        Map<String, String> dateOptions = new HashMap<String, String>();
-        dateOptions.put("java8", "Java 8 native");
-        dateOptions.put("joda", "Joda");
-        dateLibrary.setEnum(dateOptions);
-
-        cliOptions.add(dateLibrary);
 
         CliOption library = new CliOption(CodegenConstants.LIBRARY, "library template (sub-template) to use");
         library.setDefault(DEFAULT_LIBRARY);
@@ -108,32 +98,6 @@ public class JaxRSServerCodegen extends JavaClientCodegen implements CodegenConf
         if (additionalProperties.containsKey("dateLibrary")) {
             setDateLibrary(additionalProperties.get("dateLibrary").toString());
             additionalProperties.put(dateLibrary, "true");
-        }
-
-        if("joda".equals(dateLibrary)) {
-            typeMapping.put("date", "LocalDate");
-            typeMapping.put("DateTime", "DateTime");
-
-            importMapping.put("LocalDate", "org.joda.time.LocalDate");
-            importMapping.put("DateTime", "org.joda.time.DateTime");
-
-            supportingFiles.add(new SupportingFile("JodaDateTimeProvider.mustache",
-                    (sourceFolder + '/' + apiPackage).replace(".", "/"), "JodaDateTimeProvider.java"));
-            supportingFiles.add(new SupportingFile("JodaLocalDateProvider.mustache",
-                    (sourceFolder + '/' + apiPackage).replace(".", "/"), "JodaLocalDateProvider.java"));
-        }
-        else if ("java8".equals(dateLibrary)) {
-            additionalProperties.put("java8", "true");
-            additionalProperties.put("javaVersion", "1.8");
-            typeMapping.put("date", "LocalDate");
-            typeMapping.put("DateTime", "LocalDateTime");
-            importMapping.put("LocalDate", "java.time.LocalDate");
-            importMapping.put("LocalDateTime", "java.time.LocalDateTime");
-
-            supportingFiles.add(new SupportingFile("LocalDateTimeProvider.mustache",
-                    (sourceFolder + '/' + apiPackage).replace(".", "/"), "LocalDateTimeProvider.java"));
-            supportingFiles.add(new SupportingFile("LocalDateProvider.mustache",
-                    (sourceFolder + '/' + apiPackage).replace(".", "/"), "LocalDateProvider.java"));
         }
     }
 
@@ -290,9 +254,5 @@ public class JaxRSServerCodegen extends JavaClientCodegen implements CodegenConf
     @Override
     public boolean shouldOverwrite(String filename) {
         return super.shouldOverwrite(filename) && !filename.endsWith("ServiceImpl.java") && !filename.endsWith("ServiceFactory.java");
-    }
-
-    public void setDateLibrary(String library) {
-        this.dateLibrary = library;
     }
 }
